requirements.txt
# Core dependencies
# Core dependencies
crewai[tools]>=0.140.0,<1.0.0
fpdf2>=2.8.3
streamlit>=1.46.1

# Enhanced dependencies for improved functionality
plotly>=5.17.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.0.0
colorama>=0.4.6

# Performance and caching
redis>=4.6.0
diskcache>=5.6.0

# Data validation and type hints
typing-extensions>=4.7.0
mypy>=1.5.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# API and HTTP
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0

# Data processing
numpy>=1.24.0
scipy>=1.11.0

# Configuration management
pyyaml>=6.0
toml>=0.10.0

# File handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Email functionality
smtplib-ssl>=1.0.0
email-validator>=2.0.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2023.3

# JSON handling
ujson>=5.8.0

# Environment management
python-decouple>=3.8

# Async support
asyncio>=3.4.3
aiofiles>=23.2.0

# Monitoring and metrics
psutil>=5.9.0
memory-profiler>=0.61.0

# Database support (for future use)
sqlalchemy>=2.0.0
alembic>=1.11.0

# Web scraping (for enhanced data collection)
beautifulsoup4>=4.12.0
selenium>=4.11.0

# Image processing (for enhanced PDFs)
Pillow>=10.0.0

# Compression
zipfile38>=0.0.3

# Networking
urllib3>=2.0.0

# System information
platformdirs>=3.0.0

# Enhanced dependencies for improved functionality
plotly>=5.17.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and monitoring
structlog>=23.0.0
colorama>=0.4.6

# Performance and caching
redis>=4.6.0
diskcache>=5.6.0

# Data validation and type hints
typing-extensions>=4.7.0
mypy>=1.5.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# API and HTTP
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0

# Data processing
numpy>=1.24.0
scipy>=1.11.0

# Configuration management
pyyaml>=6.0
toml>=0.10.0

# File handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Email functionality
smtplib-ssl>=1.0.0
email-validator>=2.0.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2023.3

# JSON handling
ujson>=5.8.0

# Environment management
python-decouple>=3.8

# Async support
asyncio>=3.4.3
aiofiles>=23.2.0

# Monitoring and metrics
psutil>=5.9.0
memory-profiler>=0.61.0

# Database support (for future use)
sqlalchemy>=2.0.0
alembic>=1.11.0

# Web scraping (for enhanced data collection)
beautifulsoup4>=4.12.0
selenium>=4.11.0

# Image processing (for enhanced PDFs)
Pillow>=10.0.0

# Compression
zipfile38>=0.0.3

# Networking
urllib3>=2.0.0

# System information
platform>=1.0.8






